requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.2.1
Deprecated==1.2.18
ecdsa==0.19.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-Limiter==3.12
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==5.21.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
Pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
rate-limiter==0.0.3
redis==7.1.0
rich==13.9.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
wrapt==1.17.3
